Finish asset cycling

+++ game/code/game_render_group.h
@@ -28,10 +28,10 @@ struct loaded_bitmap
     void *Memory;
     v2 AlignPercentage;
     r32 WidthOverHeight;
-    s16 Width;
-    s16 Height;
+    s32 Width;
+    s32 Height;
     // TODO: Get rid of the pitch.
-    s16 Pitch;
+    s32 Pitch;
 };
 
 struct environment_map
